backend/requirements.txt
Flask
flask-cors
flask_sqlalchemy
psycopg2-binary
Flask-Migrate
Flask-Script
marshmallow
Werkzeug